create a migration
Have not run it, its just to show how migration is created. Important thing to note is the ormconfig file, as migration classes are separate from our NestJS source code. This is because their lifecycle is maintained by typeorm cli. Since migration are outside of Nest, we cannot use the Nest feature like dependency injection etc.