make a coffee rating module and import coffeservice in it
Notice how we imported only the CoffeeModule in the CoffeeRating module, and did not list the coffeeservice as dependency. If we would have listed the coffeeservice as dependency we would have to to resolve other errors as well due to incorrect mentioning of coffesservice in the providers array. Try it to see the issue