creating our first e2e test
Notice that we are creating the e2e test for AppModule, that means we are instantiating the entire application that includes all controllers, providers, enhancers, establishing database connections, configurations and so on and so forth. Nest modules typically represent the features, so ideally we should be able to test these features in isolation independently.